requirements.txt
numpy==1.26.1
psycopg2-binary==2.9.9
